Removed api/index
Moved all initializers to respective API's index file
Updated all apps to use the new initializers
