+++ src/apps/trade/api.ts
@@ -1,1 +1,5 @@
-export { web3 } from 'api'
+import { createWeb3Api } from 'api/web3'
+
+const web3 = createWeb3Api()
+
+export { web3 }
